New system of units
It defines appropriate types for each dimensionality. These types not
only check at runtime that the units used for instantiating the object
are correct for the dimensionality, but also can catch typyng errors in
the pass of parameters or dataclass fields.

TODO: revise the code when pint fixes the problem of Quantity class
whose type is incorrectly inferred as Any by mypy.

See: https://github.com/hgrecco/pint/issues/1679

When this is fixed in pint, the class
cloudmodel.unified.units.CheckedDimensionality should inherit from
Quantity, and this will remove the need of all `cast()` in the code