Solved typing issues in units
Made `CheckedDimensionality` inherit from
`pint.facects.plain.PlainQuantity` instead of inheriting from nothing
(`object`), which caused mypy to complain about any use of Quantity
methods such as `.to()` or arithmetic operations between quantities.

Making `CheckedDimensionality` inherit from  `pint.Quantity` didn't
work either because `pint.Quantity` is a dynamic class and not a valid
type for `mypy`, so it inferred `Any` and did not perform any type check
in quantities passed as parameters or assigned to attributes.

The current implementation works correctly with mypy, at the cost of
having to redefine the method `.to()` in all derived classes (Time,
Storage, etc.) to declare the proprer return type.